Unexport the ParseTextWebVTT and use internal test
