+++ webvtt.go
@@ -197,7 +197,7 @@ func ReadFromWebVTT(i io.Reader) (o *Subtitles, err error) {
 				// TODO Do something with the style
 			case webvttBlockNameText:
 				// Voice Tag to extract Speaker Name
-				item.Lines = append(item.Lines, ParseTextWebVTT(line))
+				item.Lines = append(item.Lines, parseTextWebVTT(line))
 			default:
 				// This is the ID
 				index, _ = strconv.Atoi(line)
@@ -207,8 +207,8 @@ func ReadFromWebVTT(i io.Reader) (o *Subtitles, err error) {
 	return
 }
 
-// ParseTextWebVTT parses the input line to fill the Line
-func ParseTextWebVTT(i string) (o Line) {
+// parseTextWebVTT parses the input line to fill the Line
+func parseTextWebVTT(i string) (o Line) {
 	tr := html.NewTokenizer(strings.NewReader(i))
 	var voiceName, endTag, text string
 	for {
