+++ webvtt_test.go
@@ -79,32 +79,3 @@ func TestWebVTTWithVoiceName(t *testing.T) {
 	assert.Equal(t, voiceName3, s.Items[2].Lines[0].VoiceName)
 	assert.Equal(t, "", s.Items[3].Lines[0].VoiceName)
 }
-
-func TestParseTextWebVTT(t *testing.T) {
-
-	t.Run("When both voice tags are available", func(t *testing.T) {
-		testData := `<v Bob>Correct tag</v>`
-
-		s := astisub.ParseTextWebVTT(testData)
-		assert.Equal(t, "Bob", s.VoiceName)
-		assert.Equal(t, 1, len(s.Items))
-		assert.Equal(t, "Correct tag", s.Items[0].Text)
-	})
-
-	t.Run("When there is no end tag", func(t *testing.T) {
-		testData := `<v Bob> Text without end tag`
-
-		s := astisub.ParseTextWebVTT(testData)
-		assert.Equal(t, "Bob", s.VoiceName)
-		assert.Equal(t, 1, len(s.Items))
-		assert.Equal(t, "Text without end tag", s.Items[0].Text)
-	})
-
-	t.Run("When the end tag is correct", func(t *testing.T) {
-		testData := `<v Bob>Incorrect end tag</vi>`
-
-		s := astisub.ParseTextWebVTT(testData)
-		assert.Equal(t, "", s.VoiceName)
-		assert.Equal(t, 0, len(s.Items))
-	})
-}
